Alpha0.02

+++ Server/Service/DataBases/TModerateProducts.cs
@@ -12,8 +12,8 @@ namespace Service
 
         public int Id { get; set; }
         public string Name { get; set; }
-        public int? WholesalePrice { get; set; }
-        public int RetailPrice { get; set; }
+        public double? WholesalePrice { get; set; }
+        public double RetailPrice { get; set; }
         public string Description { get; set; }
         public DateTime RequestDate { get; set; }
         public int Iddeveloper { get; set; }
