+++ Server/Service/DataBases/TProducts.cs
@@ -17,8 +17,8 @@ namespace Service
         public int Id { get; set; }
         public string Name { get; set; }
         public double? Rate { get; set; }
-        public int? WholesalePrice { get; set; }
-        public int RetailPrice { get; set; }
+        public double? WholesalePrice { get; set; }
+        public double RetailPrice { get; set; }
         public string Description { get; set; }
         public DateTime ReleaseDate { get; set; }
         public int? Quantity { get; set; }
